# pkg/webapi/docs/gen/swagger.yaml
basePath: /api
definitions:
...
  dto.MessageChange:
    properties:
      id:
        type: integer
      kind:
        type: string
      message:
        $ref: '#/definitions/dto.MessageResponse'
    type: object
  dto.MessageListResponse:
    properties:
      changes:
        items:
          $ref: '#/definitions/dto.MessageChange'
        type: array
      cursor:
        type: string
    type: object
...
  dto.MessageResponse:
    properties:
      acked_at:
        type: string
      attempts:
        type: integer
      channel:
        type: integer
      created_at:
        type: string
      direction:
        description: '"in" | "out"'
        type: string
      extended:
        description: |-
          Extended is true when the transmitted body exceeded the default
          MaxMessageText (67). The UI renders an "extended" badge on these
          rows so operators can correlate if recipients report missing or
          truncated messages. Derived from len(Text) > MaxMessageText; no
          dedicated column.
        type: boolean
      failure_reason:
        type: string
      from_call:
        type: string
      id:
        type: integer
      invite_accepted_at:
        description: |-
          InviteAcceptedAt is audit-only: set when the local operator
          accepted this invite. The UI must NOT use this to decide "joined"
          state — that comes from the live TacticalSet cache. Kept so
          operators can see when/if an invite was acted on.
        type: string
      invite_tactical:
        description: |-
          InviteTactical is the tactical callsign referenced by an invite.
          Empty (and omitted) on non-invite rows.
        type: string
      is_ack:
        type: boolean
      is_bulletin:
        type: boolean
      kind:
        description: |-
          Kind is the body classification. Always populated — "text" for
          normal messages, "invite" for tactical invitations. Never omitted
          so clients can use a simple equality check without worrying about
          a legacy empty string.
        type: string
      msg_id:
        type: string
      next_retry_at:
        type: string
      our_call:
        type: string
      path:
        type: string
      peer_call:
        type: string
      received_at:
        type: string
      received_by_call:
        type: string
      sent_at:
        type: string
      source:
        description: '"rf" | "is"'
        type: string
      status:
        description: derived — see DeriveMessageStatus
        type: string
      text:
        type: string
      thread_key:
        type: string
      thread_kind:
        type: string
      to_call:
        type: string
      unread:
        type: boolean
      via:
        type: string
    type: object
...
  dto.SendMessageRequest:
    properties:
      channel:
        description: Channel overrides the configured TX channel. Nil = use default.
        type: integer
      client_id:
        description: |-
          ClientID is an opaque client-side correlation token. Echoed back
          unchanged in the response so the optimistic UI can reconcile its
          local row with the persisted ID.
        type: string
      invite_tactical:
        description: |-
          InviteTactical is the tactical callsign referenced by an invite.
          Must be set when Kind == "invite"; ignored otherwise.
        type: string
      kind:
        description: |-
          Kind classifies the outbound row. Empty or "text" is a normal
          DM/tactical message; "invite" makes the sender build a
          `!GW1 INVITE <InviteTactical>` body and stamp the row with
          Kind=invite + InviteTactical. The sender (Phase 2) is
          responsible for honoring this; the DTO just carries it.
        type: string
      path:
        description: |-
          Path overrides the default RF path from preferences. Empty =
          use MessagePreferences.DefaultPath.
        type: string
      prefer_is:
        description: |-
          PreferIS, when true, routes the outbound via APRS-IS regardless
          of the current fallback policy.
        type: boolean
      text:
        description: |-
          Text is the message body (<= 67 APRS chars after validation).
          Ignored when Kind == "invite" — the server builds the wire body
          from InviteTactical.
        type: string
      to:
        description: |-
          To is the addressee: a station callsign for a DM or a tactical
          label for a group broadcast. Uppercase-normalized server-side.
        type: string
    type: object
...
  webapi.StatusChannel:
    properties:
      audio_peak:
        type: number
      bit_rate:
        type: integer
      dcd_state:
        type: boolean
      device_clipping:
        type: boolean
      device_peak_dbfs:
        type: number
      device_rms_dbfs:
        type: number
      id:
        type: integer
      input_device_id:
        type: integer
      modem_type:
        type: string
      name:
        type: string
      rx_bad_fcs:
        type: integer
      rx_frames:
        type: integer
      tx_frames:
        type: integer
    type: object
  webapi.StatusDTO:
    properties:
      channels:
        description: Channels is the per-channel live status snapshot (frame counters, audio levels).
        items:
          $ref: '#/definitions/webapi.StatusChannel'
        type: array
      igate:
        allOf:
          - $ref: '#/definitions/webapi.StatusIgateDTO'
        description: Igate is the current iGate session state; omitted when no iGate is configured.
      uptime_seconds:
        description: UptimeSeconds is the server process uptime in whole seconds.
        type: integer
    type: object
  webapi.StatusIgateDTO:
    properties:
      callsign:
        description: Callsign is the login callsign-SSID presented to APRS-IS.
        type: string
      connected:
        description: Connected is true while an APRS-IS session is established.
        type: boolean
      is_to_rf_gated:
        description: Downlinked is the cumulative count of IS packets transmitted on RF.
        type: integer
      last_connected:
        description: LastConnected is the UTC RFC3339 timestamp of the most recent successful IS login; omitted if never connected.
        type: string
      packets_filtered:
        description: Filtered is the cumulative count of packets dropped by the filter engine.
        type: integer
      rf_to_is_dropped:
        description: DroppedOffline is the cumulative count of RF packets dropped because the IS session was offline.
        type: integer
      rf_to_is_gated:
        description: Gated is the cumulative count of RF packets forwarded to APRS-IS.
        type: integer
      server:
        description: Server is the APRS-IS host:port currently in use (or last attempted).
        type: string
      simulation_mode:
        description: SimulationMode is true when RF->IS uploads are suppressed for testing.
        type: boolean
    type: object
...
  webtypes.ErrorResponse:
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
  description: REST API for graywolf configuration and control.
  title: Graywolf Management API
  version: "1.0"
...
paths:
...
  /messages:
    get:
      operationId: listMessages
      parameters:
        - description: 'Folder filter: inbox|sent|all'
          in: query
          name: folder
          type: string
        - description: PeerCall filter
          in: query
          name: peer
          type: string
        - description: dm|tactical
          in: query
          name: thread_kind
          type: string
        - description: Exact thread key (peer callsign for DM, tactical label for tactical)
          in: query
          name: thread_key
          type: string
        - description: Only messages at or after this RFC3339 timestamp
          in: query
          name: since
          type: string
        - description: Opaque cursor from a prior response; pages forward
          in: query
          name: cursor
          type: string
        - description: Restrict to unread rows
          in: query
          name: unread_only
          type: boolean
        - description: Cap result count (1..500)
          in: query
          name: limit
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/webtypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/webtypes.ErrorResponse'
      security:
        - CookieAuth: []
      summary: List messages
      tags:
        - messages
    post:
      consumes:
        - application/json
      operationId: sendMessage
      parameters:
        - description: Compose payload
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/dto.SendMessageRequest'
      produces:
        - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/webtypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/webtypes.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/webtypes.ErrorResponse'
      security:
        - CookieAuth: []
      summary: Send message
      tags:
        - messages
...
  /status:
    get:
      operationId: getStatus
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/webapi.StatusDTO'
      security:
        - CookieAuth: []
      summary: System status dashboard
      tags:
        - status
...
schemes:
  - http
securityDefinitions:
  CookieAuth:
    description: |-
      Session cookie issued by POST /api/auth/login.
      Swagger 2.0 lacks a native `in: cookie` apiKey
      location, so the spec models the same credential
      as a `Cookie:` request header. Browsers send this
      automatically once the session cookie is set; the
      session cookie is named `graywolf_session`.
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"